Refactor school bloc packages
