+++ lib/features/list_art/presentation/bloc/auth/school_auth_state.dart
