+++ lib/features/list_art/presentation/bloc/gallery/school_gallery_event.dart
@@ -1,0 +1,5 @@
+import 'package:equatable/equatable.dart';
+
+abstract class SchoolGalleryEvent extends Equatable {
+  const SchoolGalleryEvent();
+}
