+++ lib/features/list_art/presentation/bloc/gallery/school_gallery_state.dart
@@ -1,0 +1,10 @@
+import 'package:equatable/equatable.dart';
+
+abstract class SchoolGalleryState extends Equatable {
+  const SchoolGalleryState();
+}
+
+class InitialSchoolGalleryState extends SchoolGalleryState {
+  @override
+  List<Object> get props => [];
+}
