+++ lib/features/list_art/presentation/page/registration_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:student_art_collection/core/presentation/widget/empty_container.dart';
-import 'package:student_art_collection/features/list_art/presentation/bloc/school_auth_bloc.dart';
-import 'package:student_art_collection/features/list_art/presentation/bloc/school_auth_event.dart';
-import 'package:student_art_collection/features/list_art/presentation/bloc/school_auth_state.dart';
+import 'package:student_art_collection/features/list_art/presentation/bloc/auth/school_auth_bloc.dart';
+import 'package:student_art_collection/features/list_art/presentation/bloc/auth/school_auth_event.dart';
+import 'package:student_art_collection/features/list_art/presentation/bloc/auth/school_auth_state.dart';
 import 'package:student_art_collection/features/list_art/presentation/page/school_gallery_page.dart';
 import 'package:student_art_collection/features/list_art/presentation/widget/auth_input_decoration.dart';
 import 'package:student_art_collection/features/list_art/presentation/widget/horizontal_progress_bar.dart';
